Update interpro_processor.py
Major update for tsv file output, the domains distribution is now returned in the right order according to domain position in the sequence. It's just a classification reframing  and does not change fasta output results. If entry A and B are provided for a given sequence:  A:d2 A:d3 ; B:d1 B:d4 - the order in the TSV will be B:d1, B:d4, A:d2 A:d3. This will rarely happens. And won't impact any trimming process. It's just decoration for the tsv file to be more  readable. Otherwise, we must attribute to  each d1,d2,d3 a specific entry. Which may be unpleasant to read. Also, the advantage is to clearly see which entry compete with another for a specific domain position.